+++ Tests/3. Integration Tests/PSGSuite.Integration.Tests.ps1
@@ -1,21 +1,22 @@
-$projectRoot = Resolve-Path "$PSScriptRoot\..\.."
-$ModulePath = Resolve-Path "$projectRoot\BuildOutput\$($env:BHProjectName)"
-$decompiledModulePath = Resolve-Path "$projectRoot\$($env:BHProjectName)"
+if ($ENV:BUILD_BUILDURI -like 'vstfs:*') {
+    $projectRoot = Resolve-Path "$PSScriptRoot\..\.."
+    $ModulePath = Resolve-Path "$projectRoot\BuildOutput\$($env:BHProjectName)"
+    $decompiledModulePath = Resolve-Path "$projectRoot\$($env:BHProjectName)"
 
-$Verbose = @{}
-if ($ENV:BHBranchName -eq "development" -or $env:BHCommitMessage -match "!verbose") {
-    $Verbose.add("Verbose",$True)
+    $Verbose = @{}
+    if ($ENV:BHBranchName -eq "development" -or $env:BHCommitMessage -match "!verbose") {
+        $Verbose.add("Verbose",$True)
+    }
+    $moduleRoot = Split-Path (Resolve-Path "$ModulePath\*\*.psd1")
+
+    Import-Module $ModulePath -Force -Verbose:$false
+    Import-PSGSuiteConfig -Json $env:PSGSuiteConfigJson -Temporary -Verbose
+
+    $u = Get-GSUser
+    $u | Select-Object @{N="GivenName";E={$_.Name.GivenName}},OrgUnitPath,Kind
+
+    Send-GmailMessage -From $u.PrimaryEmail -To $u.PrimaryEmail -Subject "Hello from Azure Pipelines + PS Version $($PSVersionTable.PSVersion.ToString())!" -Body "<pre>`n$((Get-ChildItem Env: | Where-Object {$_.Name -match '^(BUILD_|BH).*$'} | Format-Table -AutoSize | Out-String).Trim())`n</pre>" -BodyAsHtml -Verbose
 }
-$moduleRoot = Split-Path (Resolve-Path "$ModulePath\*\*.psd1")
-
-Import-Module $ModulePath -Force -Verbose:$false
-Import-PSGSuiteConfig -Json $env:PSGSuiteConfigJson -Temporary -Verbose
-
-$u = Get-GSUser
-$u | Select-Object @{N="GivenName";E={$_.Name.GivenName}},OrgUnitPath,Kind
-
-Send-GmailMessage -From $u.PrimaryEmail -To $u.PrimaryEmail -Subject "Hello from Azure Pipelines + PS Version $($PSVersionTable.PSVersion.ToString())!" -Body "<pre>`n$((Get-ChildItem Env: | Where-Object {$_.Name -match '^(BUILD_|BH).*$'} | Format-Table -AutoSize | Out-String).Trim())`n</pre>" -BodyAsHtml -Verbose
-
 <#
 Describe "Function contents" -Tag 'Module' {
     Context "Get-GSUser should return a user" {